feat: Expose a sink adapter for interfacing custom sinks
This adapter accepts a writer which is callable and accepts the log
entry as parameter. This allows the user to define custom behaviours
while keeping the same sink interface as the other ones.

A straightforward example is the NvimNotify sink that adapts the log
entry into compatible arguments for the plugin.

 - Adapt the NvimNotify sink to make use of the new adapter class
 - Add + update existing tests